requirements.txt
python-dotenv==1.0.0
retell-sdk==1.0.0
openai==1.3.5
deepgram-sdk==2.11.0
elevenlabs==0.2.24
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pydantic==2.5.2
python-multipart==0.0.6
PyYAML==6.0.1
sounddevice==0.4.6
numpy==1.26.2
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
loguru==0.7.2
